add selectorElement class
