+++ Composer/packages/extensions/visual-designer/src/constants/ElementAttributes.ts
@@ -14,29 +14,3 @@ export enum AttrNames {
   FocusableElement = 'data-is-focusable',
   SelectionIndex = 'data-selection-index',
 }
-
-export class AbstractSelectorElement {
-  bounds: DOMRect;
-  [AttrNames.SelectableElement]: string | undefined;
-  [AttrNames.NodeElement]: string | undefined;
-  [AttrNames.EdgeMenuElement]: string | undefined;
-  [AttrNames.FocusedId]: string | undefined;
-  [AttrNames.SelectedId]: string | undefined;
-  [AttrNames.Tab]: string | undefined;
-
-  constructor(element: HTMLElement) {
-    this.bounds = element.getBoundingClientRect() as DOMRect;
-
-    Object.keys(AttrNames).forEach(key => {
-      this[AttrNames[key]] = element.getAttribute(AttrNames[key]);
-    });
-  }
-
-  getAttribute(attrName) {
-    return this[attrName];
-  }
-
-  getBoundingClientRect() {
-    return this.bounds;
-  }
-}
